update with new fields, fix Multiplier type

+++ IbFlexReader/IbFlexReader.Contracts/Ib/SLBActivity.cs
@@ -44,7 +44,7 @@
 
         public string Issuer { get; set; }
 
-        public int? Multiplier { get; set; }
+        public double? Multiplier { get; set; }
 
         public double? Strike { get; set; }
